refactor(broker-core): refactor wf instance processor
